requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
jupyter
sqlalchemy
pytest-cov